Fix bugs with Lambdas, Refactoring
- Fix chainId passing to /pools and /sor
- Move token related functionality to tokens.ts
- Add error handling if CoinGecko fails to return a price
- Make getPool and getToken return undefined instead of throwing an
error if they can't find the item.
